Refactor pages and app.js
