+++ client/src/pages/AdminDashboard.js
@@ -1,0 +1,9 @@
+const AdminDashboard = () => {
+  return (
+    <div>
+        <h1>AdminDashboard</h1>
+    </div>
+  )
+}
+
+export default AdminDashboard
