+++ client/src/pages/Error.js
@@ -1,0 +1,9 @@
+const Error = () => {
+  return (
+    <div>
+      <h1>Error</h1>
+    </div>
+  )
+}
+
+export default Error
