+++ client/src/pages/Register.js
@@ -1,0 +1,9 @@
+const Register = () => {
+  return (
+    <div>
+        <h1>Register</h1>
+    </div>
+  )
+}
+
+export default Register
