+++ client/src/pages/index.js
@@ -1,0 +1,13 @@
+import Landing from './Landing'
+import Register from './Register'
+import Error from './Error'
+import StudentDashboard from './StudentDashboard'
+import AdminDashboard from './AdminDashboard'
+
+export {
+    Landing,
+    Register,
+    Error,
+    StudentDashboard,
+    AdminDashboard
+}
